use fts on mac, legacy on linux
