+++ src/unix/fts.cc
@@ -1,10 +1,6 @@
 #include <string>
-
 #define __THROW // weird error on linux
-
-#include <sys/stat.h>
-#include <dirent.h>
-
+#include <fts.h>
 #include "../DirTree.hh"
 #include "../shared/BruteForceBackend.hh"
 
@@ -12,36 +8,36 @@
 #if __APPLE__
 #define st_mtim st_mtimespec
 #endif
-#define ISDOT(a) (a[0] == '.' && (!a[1] || (a[1] == '.' && !a[2])))
 
-void iterateDir(Watcher &watcher, const std::shared_ptr <DirTree> tree, const char *dirname) {
-    if (DIR *dir = opendir(dirname)) {
-        while (struct dirent *ent = (errno = 0, readdir(dir))) {
-            if (ISDOT(ent->d_name)) continue;
+void BruteForceBackend::readTree(Watcher &watcher, std::shared_ptr<DirTree> tree) {
+  char *paths[2] {(char *)watcher.mDir.c_str(), NULL};
+  FTS *fts = fts_open(paths, FTS_NOCHDIR | FTS_PHYSICAL, NULL);
+  if (!fts) {
+    throw WatcherError(strerror(errno), &watcher);
+  }
 
-            std::string fullPath = dirname + std::string("/") + ent->d_name;
+  FTSENT *node;
+  bool isRoot = true;
 
-            if (watcher.mIgnore.count(fullPath) == 0) {
-                struct stat attrib;
-                stat(fullPath.c_str(), &attrib);
-                bool isDir = ent->d_type == DT_DIR;
-
-                tree->add(fullPath, CONVERT_TIME(attrib.st_mtim), isDir);
-
-                if (isDir) {
-                    iterateDir(watcher, tree, fullPath.c_str());
-                }
-            }
-        }
-
-        closedir(dir);
+  while ((node = fts_read(fts)) != NULL) {
+    if (node->fts_errno) {
+      fts_close(fts);
+      throw WatcherError(strerror(node->fts_errno), &watcher);
     }
 
-    if (errno) {
-        throw WatcherError(strerror(errno), &watcher);
+    if (isRoot && !(node->fts_info & FTS_D)) {
+      fts_close(fts);
+      throw WatcherError(strerror(ENOTDIR), &watcher);
     }
-}
 
-void BruteForceBackend::readTree(Watcher &watcher, std::shared_ptr <DirTree> tree) {
-    return iterateDir(watcher, tree, watcher.mDir.c_str());
+    if (watcher.mIgnore.count(std::string(node->fts_path)) > 0) {
+      fts_set(fts, node, FTS_SKIP);
+      continue;
+    }
+
+    tree->add(node->fts_path, CONVERT_TIME(node->fts_statp->st_mtim), (node->fts_info & FTS_D) == FTS_D);
+    isRoot = false;
+  }
+
+  fts_close(fts);
 }
